requirements.txt
google-generativeai
langchain
pymupdf
flask
mecab-python3
jieba
langdetect
